make MetricsMock thread safe
