+++ filters/apiusagemonitoring/filter_consumermetrics_test.go
@@ -9,11 +9,6 @@ package apiusagemonitoring
 
 import (
 	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"github.com/stretchr/testify/assert"
-	"github.com/zalando/skipper/filters"
-	"github.com/zalando/skipper/metrics/metricstest"
 	"net/http"
 	"testing"
 )
@@ -341,124 +336,3 @@ func Test_Filter_ClientMetrics_NoFlagClientIdKeyName(t *testing.T) {
 //		expectingNoClientBasedMetrics: true,
 //	})
 //}
-
-func testClientMetrics(
-	t *testing.T,
-	testCase clientMetricsTest,
-) {
-	conf := testWithFilterConf{
-		url:    testCase.url,
-		header: testCase.header,
-		filterCreate: func() (filters.Filter, error) {
-			filterConf := map[string]interface{}{
-				"application_id": "my_app",
-				"api_id":         "my_api",
-				"path_templates": []string{
-					"foo/orders",
-				},
-			}
-			if testCase.clientTrackingPattern != "" {
-				filterConf["client_tracking_pattern"] = testCase.clientTrackingPattern
-			}
-			js, err := json.Marshal(filterConf)
-			if !assert.NoError(t, err) {
-				t.FailNow()
-			}
-			args := []interface{}{string(js)}
-			spec := NewApiUsageMonitoring(true, testCase.realmKeyName, testCase.clientIdKeyName)
-			return spec.CreateFilter(args)
-		},
-	}
-	previousLatencySum := float64(0)
-	testWithFilterC(t, conf, func(t *testing.T, pass int, m *metricstest.MockMetrics) {
-
-		expectedCounters := make([]string, 0)
-		expectedFloatCounters := make([]string, 0)
-		expectedMeasures := make([]string, 0)
-
-		//
-		// Assert consumer metrics
-		//
-		if testCase.expectedConsumerMetricPrefix != "" {
-
-			httpCountKey := testCase.expectedConsumerMetricPrefix + "http_count"
-			expectedCounters = append(expectedCounters, httpCountKey)
-			if assert.Contains(t, m.Counters, httpCountKey) {
-				v := m.Counters[httpCountKey]
-				assert.Equal(t, int64(pass), v)
-			}
-
-			httpClassCountKey := testCase.expectedConsumerMetricPrefix + "http2xx_count"
-			expectedCounters = append(expectedCounters, httpClassCountKey)
-			if assert.Contains(t, m.Counters, httpClassCountKey) {
-				v := m.Counters[httpClassCountKey]
-				assert.Equal(t, int64(pass), v)
-			}
-
-			latencySumKey := testCase.expectedConsumerMetricPrefix + "latency_sum"
-			expectedFloatCounters = append(expectedFloatCounters, latencySumKey)
-			if assert.Contains(t, m.FloatCounters, latencySumKey) {
-				v := m.FloatCounters[latencySumKey]
-				assert.Conditionf(t,
-					func() bool {
-						return v > previousLatencySum
-					}, "current consumer latency sum is not higher than the previous recorded one (%f to %f)",
-					previousLatencySum, v)
-			}
-
-		}
-
-		//
-		// Assert endpoint metrics
-		//
-		if testCase.expectedEndpointMetricPrefix != "" {
-
-			httpCountKey := testCase.expectedEndpointMetricPrefix + "http_count"
-			expectedCounters = append(expectedCounters, httpCountKey)
-			if assert.Contains(t, m.Counters, httpCountKey) {
-				v := m.Counters[httpCountKey]
-				assert.Equal(t, int64(pass), v)
-			}
-
-			httpCountClassKey :=testCase.expectedEndpointMetricPrefix+"http2xx_count"
-			expectedCounters = append(expectedCounters, httpCountClassKey)
-			if assert.Contains(t,  m.Counters, httpCountClassKey) {
-				v := m.Counters[httpCountClassKey]
-				assert.Equal(t, int64(pass), v)
-			}
-
-			latencyKey := testCase.expectedEndpointMetricPrefix + "latency"
-			expectedMeasures = append(expectedMeasures, latencyKey)
-			assert.Contains(t, m.Measures, latencyKey)
-		}
-
-		actualCounters := make([]string, 0, len(m.Counters))
-		for k := range m.Counters {
-			actualCounters = append(actualCounters, k)
-		}
-		assert.ElementsMatch(t, expectedCounters, actualCounters)
-
-		actualFloatCounters := make([]string, 0, len(m.FloatCounters))
-		for k := range m.FloatCounters {
-			actualFloatCounters = append(actualFloatCounters, k)
-		}
-		assert.ElementsMatch(t, expectedFloatCounters, actualFloatCounters)
-
-		actualMeasures := make([]string, 0, len(m.Measures))
-		for k := range m.Measures {
-			actualMeasures = append(actualMeasures, k)
-		}
-		assert.ElementsMatch(t, expectedMeasures, actualMeasures)
-
-	})
-}
-
-func buildFakeJwtWithBody(jwtBodyJson map[string]interface{}) string {
-	jwtBodyBytes, err := json.Marshal(jwtBodyJson)
-	if err != nil {
-		panic(err)
-	}
-	jwtBody := base64.RawURLEncoding.EncodeToString(jwtBodyBytes)
-	jwt := fmt.Sprintf("<No Header>.%s.< No Verify Signature>", jwtBody)
-	return jwt
-}
